+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.8
 
 - Improved parser builder `BinaryExpression`. The generated parser is now smaller, faster and uses less memory (does not use list and tuples)
+- Improved parser builder `IdentifierExpression`. The generated parser is now faster (using binary search)
 
 ## 1.0.7